Midway through changes.
Will complete later.
