+++ cluster_gas/main.py
@@ -1,31 +1,14 @@
 import numpy as np
-from lib import ODE_int, generate_matrix, parameters as pr
+from lib import cluster_system
 import matplotlib.pyplot as plt
 
 N_ITER = 5000
-n_class = 10
-C_init = np.zeros(shape=(n_class+1))
-C_init[0] = 1.0
-
-
-def calc_mass(x):
-    return sum((index+1)*value for index, value in enumerate(x))
-
-def solve_system(n_class=n_class):
-    soln = [C_init]
-    times = np.linspace(0, 0.8e-12, N_ITER)
-    
-    for t in times:
-        M = generate_matrix.update_matrix(n_class=n_class, C_1=soln[-1][0])
-        #print M
-        x_new = ODE_int.RK4(M, soln[-1][0:-1], t, 1)[0]
-        #print x_new
-        #x_new[0] += (1-calc_mass(x_new))  # Add this line to conserve mass?
-        soln.append(x_new)
-    
-    return np.asarray(soln)
 
 if __name__ == "__main__":
+
+    sys1 = cluster_system.ClusterSystem( # insert args...)
+
+
     soln = solve_system()
     print soln[-1][:-1]
 
@@ -56,10 +39,3 @@ if __name__ == "__main__":
     
     #print b_1 * soln[-1][0]**2, pr.alpha(2) * soln[-1][1]
     #print b_1 * soln[-1][0] * soln[-1][1], pr.alpha(3) * soln[-1][2]
-    
-    
-    
-    
-    
-    
-
